Add documentation for Mirage
After Slack-pairing session with Dan Freeman, it turns out that if one wants to emulate failures with `mirage` w/o creating new models, `mirage` will not call `serialize` hook which `ember-cli-pact` uses to record request and later on write them into a file. The workaround is to call `recordRequest` manually so the interaction is recorded.